[ResourceTiming]: Don't expose nextHopProtocol when TAO fails
The nextHopProtocol attribute was originally special-cased to be exposed
for 'non-document' destinations even if the Timing-Allow-Origin check
failed. This CL removes the special case so that nextHopProtocol behaves
like the other potentially-sensitive fields of PerformanceResourceTiming
entries; the protocol string will only be exposed for same-origin
requests or when the Timing-Allow-Origin header matches the requester's
origin.

BUG=1184726

Change-Id: I5d758a2263373636fdf23be97bdc642489c0cf93