Move COEP to Policy Container
In order to properly support inheritance of COEP within blob, move COEP
in policy container. This also consolidates its inheritance within the
policy container.
Remains specific case where inheritance is handled only for COEP but
not PolicyContainer: Extensions and fenced frames [1]. This was moved
within `void NavigationRequest::ComputePoliciesToCommit()`.
COEP enforcement was added in NavigationRequest::BeginNavigationImpl,
in the cases where an URL loader is not needed. This ensures COEP is
enforced on non-initial about:blank.

[1] https://docs.google.com/document/d/11ICO7dKEjeeJtSHyVitcN5JiOD0uRgvRzyyo3FY0y1c/edit?pli=1&resourcekey=0-bTO8Ltbpfi05Xt4aK-MeEQ#
bug: 1151223
Change-Id: I33a0dd61c257b6927fa335bdd983fb1b393c6177