Add tests for :link or :visited in :has()
Added unit tests and web platform tests for :link and :visited
in :has() argument to check leaking visitedness by :has().

When we have ':has(:visited) {...}', the subject elements of the rule
are the ancestors of the visited link element.

To prevent leaking visitedness to the link's ancestors, the ':visited'
selector does not match if it is inside the ':has()' argument selector.
So if a ':has()' argument selector requires a matching ':visited', the
style rule are not applied.

In order to apply the consistent limitation to ':has()' invalidation,
ignored the ':visited' pseudo class when building pseudo type hash set
for ':has()' invalidation to prevent invalidation triggered by :visited.

Bug: 669058
Change-Id: Ibc781c213e454db24fe892d3731f4f26e578a960